GVT-2479: Split all location track infoboxes into their own components
