+++ ui/src/tool-panel/location-track/location-track-infobox-linking-container.tsx
@@ -44,14 +44,6 @@ const LocationTrackInfoboxLinkingContainer: React.FC<LocationTrackInfoboxLinking
                 linkingState={trackLayoutState.linkingState}
                 splittingState={trackLayoutState.splittingState}
                 onDataChange={onDataChange}
-                onStartLocationTrackGeometryChange={(interval) => {
-                    delegates.showLayers(['alignment-linking-layer']);
-                    delegates.startAlignmentGeometryChange(interval);
-                }}
-                onEndLocationTrackGeometryChange={() => {
-                    delegates.hideLayers(['alignment-linking-layer']);
-                    delegates.stopLinking();
-                }}
                 showArea={delegates.showArea}
                 publishType={trackLayoutState.publishType}
                 changeTimes={changeTimes}
@@ -65,7 +57,6 @@ const LocationTrackInfoboxLinkingContainer: React.FC<LocationTrackInfoboxLinking
                     trackLayoutState.map.verticalGeometryDiagramState.visible
                 }
                 onHighlightItem={onHoverOverPlanSection}
-                showLocationTrackTaskList={delegates.showLocationTrackTaskList}
             />
         );
 };
